+++ factfinder/geography/2010.py
@@ -1,3 +1,4 @@
+import itertools
 import math
 from pathlib import Path
 
@@ -7,16 +8,15 @@ from cached_property import cached_property
 
 
 class AggregatedGeography:
-    def __init__(self, year=2018):
-        self.year = year
+    def __init__(self):
+        self.year = 2010
 
     @cached_property
     def lookup_geo(self):
-        year = (
-            self.year // 10 * 10
-        )  # find the current decennial year based on given year
+        # find the current decennial year based on given year
         lookup_geo = pd.read_csv(
-            f"{Path(__file__).parent}/data/lookup_geo/{year}/lookup_geo.csv",
+            Path(__file__).parent.parent
+            / f"data/lookup_geo/{self.year}/lookup_geo.csv",
             dtype="str",
         )
         lookup_geo["geoid_block"] = lookup_geo.county_fips + lookup_geo.ctcb2010
@@ -208,3 +208,12 @@ class AggregatedGeography:
                 },
             },
         }
+
+    @cached_property
+    def aggregated_geography(self) -> list:
+        list3d = [
+            [list(k.keys()) for k in i.values()]
+            for i in self.aggregate_vertical_options.values()
+        ]
+        list2d = itertools.chain.from_iterable(list3d)
+        return list(set(itertools.chain.from_iterable(list2d)))
