--- pipelines/acs/README.md
@@ -1,2 +1,0 @@
-# ACS5 pipeline for Population Factfinder
-## Instructions:
